# specs/004-ui-implementation/contracts/eligibility-wizard-ui.openapi.yaml
openapi: 3.0.3
info:
  title: Eligibility Wizard UI API
  version: 1.0.0
  description: API contracts for the Eligibility Wizard UI.
servers:
  - url: http://localhost:5008
paths:
  /api/sessions:
    post:
      summary: Create anonymous session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionDto"
      responses:
        "201":
          description: Session created
          headers:
            Set-Cookie:
              description: Session cookie (MAA_SessionId)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDto"
        "400":
          description: Invalid request
        "500":
          description: Server error
  /api/sessions/{sessionId}:
    get:
      summary: Get session by id
      operationId: getSession
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Session details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDto"
        "401":
          description: Session expired
        "404":
          description: Session not found
      security:
        - sessionCookie: []
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: MAA_SessionId
  schemas:
    CreateSessionDto:
      type: object
      properties:
        ipAddress:
          type: string
          maxLength: 45
        userAgent:
          type: string
          maxLength: 500
        timeoutMinutes:
          type: integer
          minimum: 15
          maximum: 480
        inactivityTimeoutMinutes:
          type: integer
          minimum: 5
          maximum: 60
      required:
        - ipAddress
        - userAgent
    SessionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        state:
          type: string
        userId:
          type: string
          format: uuid
          nullable: true
        ipAddress:
          type: string
        userAgent:
          type: string
        sessionType:
          type: string
        encryptionKeyVersion:
          type: integer
        expiresAt:
          type: string
          format: date-time
        inactivityTimeoutAt:
          type: string
          format: date-time
        lastActivityAt:
          type: string
          format: date-time
        isRevoked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - state
        - ipAddress
        - userAgent
        - sessionType
        - encryptionKeyVersion
        - expiresAt
        - inactivityTimeoutAt
        - lastActivityAt
        - isRevoked
        - createdAt
        - updatedAt
...